+++ CHANGELOG.md
@@ -13,6 +13,8 @@ directly impact users rather than highlighting other key architectural updates.*
 - The current balance UI on top of the Account screen has been reworked. It now displays the currently available 
   balance as well.
 - The same current balance UI was also incorporated into the Send Form screen. 
+- The Send Error screen now contains a simple text with the reason for failure. The Send screen UI refactoring is 
+  still in progress.
 
 ## [0.2.0 (530)] - 2024-01-16